+++ spec/lib/karafka/web/management/migrations/consumers_metrics/0_set_initial_spec.rb
@@ -1,8 +1,10 @@
 # frozen_string_literal: true
 
-RSpec.describe Karafka::Web::Management::Migrations::SetInitialConsumersState do
+RSpec.describe(
+  Karafka::Web::Management::Migrations::ConsumersMetrics::SetInitial
+) do
   it { expect(described_class.versions_until).to eq('0.0.1') }
-  it { expect(described_class.type).to eq(:consumers_state) }
+  it { expect(described_class.type).to eq(:consumers_metrics) }
 
   # This migration is covered in the migrator specs for initial setup
 end