+++ spec/lib/karafka/web/management/migrations/consumers_states/1700234522_introduce_waiting_spec.rb
@@ -1,8 +1,10 @@
 # frozen_string_literal: true
 
-RSpec.describe Karafka::Web::Management::Migrations::IntroduceWaitingInConsumersState do
+RSpec.describe(
+  Karafka::Web::Management::Migrations::ConsumersStates::IntroduceWaiting
+) do
   it { expect(described_class.versions_until).to eq('1.2.1') }
-  it { expect(described_class.type).to eq(:consumers_state) }
+  it { expect(described_class.type).to eq(:consumers_states) }
 
   context 'when migrating from 1.1.0' do
     let(:state) { Fixtures.consumers_states_json('v1.1.0') }