+++ spec/lib/karafka/web/management/migrations/consumers_states/1706611396_rename_lag_total_to_lag_hybrid_spec.rb
@@ -1,8 +1,10 @@
 # frozen_string_literal: true
 
-RSpec.describe Karafka::Web::Management::Migrations::RenameLagTotalToLagHybridInStates do
+RSpec.describe(
+  Karafka::Web::Management::Migrations::ConsumersStates::RenameLagTotalToLagHybrid
+) do
   it { expect(described_class.versions_until).to eq('1.3.1') }
-  it { expect(described_class.type).to eq(:consumers_state) }
+  it { expect(described_class.type).to eq(:consumers_states) }
 
   context 'when migrating from 1.3.0' do
     let(:state) { Fixtures.consumers_states_json('v1.3.0') }